+++ src/main/java/kitchenpos/ui/MenuGroupRestController.java
@@ -6,7 +6,7 @@ import java.util.stream.Collectors;
 import kitchenpos.application.MenuGroupService;
 import kitchenpos.domain.MenuGroup;
 import kitchenpos.dto.request.MenuGroupCreateRequest;
-import kitchenpos.dto.request.response.MenuGroupResponse;
+import kitchenpos.dto.response.MenuGroupResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
