+++ src/main/java/kitchenpos/ui/MenuRestController.java
@@ -6,7 +6,7 @@ import java.util.stream.Collectors;
 import kitchenpos.application.MenuService;
 import kitchenpos.domain.Menu;
 import kitchenpos.dto.request.MenuCreateRequest;
-import kitchenpos.dto.request.response.MenuResponse;
+import kitchenpos.dto.response.MenuResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
